pyproject.toml
[project]
name = "ml-proj"
version = "0.1.0"
description = "Machine learning engineering framework for training, tuning, evaluation, and inference"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "joblib>=1.4.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.34.0",
]

[project.scripts]
mlproj = "mlproj.cli:main"

[project.optional-dependencies]
deploy = [
    "flask>=3.1.3",
    "onnx>=1.20.1",
    "onnxruntime>=1.23,<1.24",
    "onnxscript>=0.6.2",
    "requests>=2.32.5",
]
legacy = [
    "bayesian-optimization>=3.2.0",
]
gpu = [
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
package = true
[tool.hatch.build.targets.wheel]
packages = ["src/mlproj"]
